Arreglo configs y cambio de mati

+++ shared/include/config_utils.h
@@ -33,8 +33,8 @@ typedef struct {
     int MARCOS_POR_CARPINCHO;
     int CANTIDAD_ENTRADAS_TLB;
     char *ALGORITMO_REEMPLAZO_TLB;
-    int RETARDO_ACIERTO_TLB;
-    int RETARDO_FALLO_TLB;
+    float RETARDO_ACIERTO_TLB;
+    float RETARDO_FALLO_TLB;
     int TAMANIO_PAGINA;
     char *PATH_DUMP_TLB;
 } t_config_memoria;
@@ -46,7 +46,7 @@ typedef struct {
     int TAMANIO_PAGINA;
     t_list *ARCHIVOS_SWAP;
     int MARCOS_MAXIMOS;
-    int RETARDO_SWAP;
+    float RETARDO_SWAP;
     char *TIPO_ASIGNACION;
 } t_config_swap;
